+++ test/workspace/git_test.rb
@@ -5,83 +5,86 @@ class WorkspaceGitTest < Minitest::Test
 
   GitBlameInfo = Runners::GitBlameInfo
 
-  def test_prepare_head_source
-    with_workspace(head: "998bc02a913e3899f3a1cd327e162dd54d489a4b",
-                   git_http_url: "https://github.com", owner: "sider", repo: "runners", pull_number: 533) do |workspace|
+  def workspace_params
+    {
+      head: "998bc02a913e3899f3a1cd327e162dd54d489a4b",
+      git_http_url: "https://github.com",
+      owner: "sider",
+      repo: "runners",
+      pull_number: 533,
+    }
+  end
+
+  def workspace_params_with_base
+    workspace_params.merge(base: "abe1cfc294c8d39de7484954bf8c3d7792fd8ad1")
+  end
+
+  def test_instance_class
+    with_workspace(**workspace_params) do |workspace|
       assert_instance_of Runners::Workspace::Git, workspace
-      mktmpdir do |dest|
-        workspace.send(:prepare_head_source, dest)
-        refute_empty dest.children
-        assert (dest / "README.md").file?
-        refute (dest / ".git").exist?
-      end
     end
   end
 
-  def test_prepare_base_source_raise_if_base_is_nil
-    with_workspace(head: "998bc02a913e3899f3a1cd327e162dd54d489a4b",
-                   git_http_url: "https://github.com", owner: "sider", repo: "runners", pull_number: 533) do |workspace|
-      assert_instance_of Runners::Workspace::Git, workspace
-      mktmpdir do |dest|
-        assert_raises(ArgumentError) do
-          workspace.send(:prepare_base_source, dest)
-        end
-      end
+  def test_prepare_head_source
+    with_workspace(**workspace_params) do |workspace|
+      workspace.send(:prepare_head_source, nil)
+
+      dest = workspace.working_dir
+      refute_empty dest.children
+      assert_path_exists dest / "README.md"
+      assert_path_exists dest / ".git"
     end
   end
 
   def test_remote_url
-    with_workspace(head: "commit",
-                   git_http_url: "https://github.example.com", owner: "foo", repo: "bar") do |workspace|
-      assert_instance_of Runners::Workspace::Git, workspace
-      assert_equal URI("https://github.example.com/foo/bar"), workspace.send(:remote_url)
+    with_workspace(**workspace_params) do |workspace|
+      assert_equal URI("https://github.com/sider/runners"), workspace.send(:remote_url)
     end
   end
 
   def test_remote_url_with_token
-    with_workspace(head: "commit",
-                   git_http_url: "https://github.com", owner: "foo", repo: "bar",
-                   git_http_userinfo: "x-access-token:v1.aaabbbcccdddeeefffgggbc06400127f248a82ac") do |workspace|
-      assert_instance_of Runners::Workspace::Git, workspace
-      assert_equal URI("https://x-access-token:v1.aaabbbcccdddeeefffgggbc06400127f248a82ac@github.com/foo/bar"), workspace.send(:remote_url)
+    params = workspace_params.merge(git_http_userinfo: "x-access-token:v1.aaabbbcccdddeeefffgggbc06400127f248a82ac")
+    with_workspace(**params) do |workspace|
+      assert_equal URI("https://x-access-token:v1.aaabbbcccdddeeefffgggbc06400127f248a82ac@github.com/sider/runners"), workspace.send(:remote_url)
     end
   end
 
   def test_git_fetch_args
-    with_workspace(head: "998bc02a913e3899f3a1cd327e162dd54d489a4b",
-                   git_http_url: "https://github.com", owner: "sider", repo: "runners", pull_number: 533) do |workspace|
-      assert_instance_of Runners::Workspace::Git, workspace
-      assert_equal %w[--quiet --no-tags --no-recurse-submodules origin
-                          +refs/heads/*:refs/remotes/origin/* +refs/pull/533/head:refs/remotes/pull/533/head], workspace.send(:git_fetch_args)
+    with_workspace(**workspace_params) do |workspace|
+      assert_equal %w[
+        --quiet --no-tags --no-recurse-submodules origin
+        +refs/heads/*:refs/remotes/origin/*
+        +refs/pull/533/head:refs/remotes/pull/533/head
+      ], workspace.send(:git_fetch_args)
     end
+  end
 
-    with_workspace(head: "998bc02a913e3899f3a1cd327e162dd54d489a4b",
-                   git_http_url: "https://github.com", owner: "sider", repo: "runners") do |workspace|
-      assert_instance_of Runners::Workspace::Git, workspace
-      assert_equal %w[--quiet --no-tags --no-recurse-submodules origin
-                          +refs/heads/*:refs/remotes/origin/*], workspace.send(:git_fetch_args)
+  def test_git_fetch_args_without_pull_number
+    params = workspace_params.tap { |p| p.delete(:pull_number) }
+    with_workspace(**params) do |workspace|
+      assert_equal %w[
+        --quiet --no-tags --no-recurse-submodules origin
+        +refs/heads/*:refs/remotes/origin/*
+      ], workspace.send(:git_fetch_args)
     end
   end
 
   def test_patches_returns_nil_because_base_is_nil
-    with_workspace(head: "998bc02a913e3899f3a1cd327e162dd54d489a4b",
-                   git_http_url: "https://github.com", owner: "sider", repo: "runners", pull_number: 533) do |workspace|
-      assert_instance_of Runners::Workspace::Git, workspace
+    with_workspace(**workspace_params) do |workspace|
       assert_nil workspace.send(:patches)
     end
   end
 
   def test_patches_returns_patches
-    with_workspace(base: "abe1cfc294c8d39de7484954bf8c3d7792fd8ad1", head: "998bc02a913e3899f3a1cd327e162dd54d489a4b",
-                   git_http_url: "https://github.com", owner: "sider", repo: "runners", pull_number: 533) do |workspace|
-      assert_instance_of Runners::Workspace::Git, workspace
+    with_workspace(**workspace_params_with_base) do |workspace|
+      workspace.send(:prepare_head_source, nil)
       assert_instance_of GitDiffParser::Patches, workspace.send(:patches)
     end
   end
 
   def test_patches_not_calling_git_diff_twice
-    with_workspace(base: "abe1cfc294c8d39de7484954bf8c3d7792fd8ad1", head: "998bc02a913e3899f3a1cd327e162dd54d489a4b",
-                   git_http_url: "https://github.com", owner: "sider", repo: "runners", pull_number: 533) do |workspace|
+    with_workspace(**workspace_params_with_base) do |workspace|
+      workspace.send(:prepare_head_source, nil)
       p1 = workspace.send(:patches)
       p2 = workspace.send(:patches)
       assert_same p1, p2
@@ -90,8 +93,8 @@ class WorkspaceGitTest < Minitest::Test
   end
 
   def test_git_blame_info
-    with_workspace(base: "abe1cfc294c8d39de7484954bf8c3d7792fd8ad1", head: "998bc02a913e3899f3a1cd327e162dd54d489a4b",
-                   git_http_url: "https://github.com", owner: "sider", repo: "runners", pull_number: 533) do |workspace|
+    with_workspace(**workspace_params_with_base) do |workspace|
+      workspace.send(:prepare_head_source, nil)
       info = workspace.range_git_blame_info('test/smokes/haml_lint/expectations.rb', 137, 140)
       assert_equal(
         [